Rename app config components
