+++ marketplace/cloudinary-assets/src/components/AppConfig/AppConfig.spec.tsx
@@ -3,7 +3,7 @@ import { render, cleanup, wait, fireEvent } from '@testing-library/react';
 
 import { AppExtensionSDK } from 'contentful-ui-extensions-sdk';
 
-import CloudinaryAppConfig from './cloudinaryAppConfig';
+import AppConfig from './AppConfig';
 
 const contentTypes = [
   {
@@ -40,22 +40,24 @@ const makeSdkMock = () => ({
 });
 
 const renderComponent = (sdk: unknown) => {
-  return render(<CloudinaryAppConfig sdk={sdk as AppExtensionSDK} />);
+  return render(<AppConfig sdk={sdk as AppExtensionSDK} />);
 };
 
-describe('CloudinaryAppConfig', () => {
+describe('AppConfig', () => {
   afterEach(cleanup);
 
   it('renders app before installation', async () => {
     const sdk = makeSdkMock();
     const { getByLabelText } = renderComponent(sdk);
     await wait(() => getByLabelText(/Cloud name/));
+
     [[/Cloud name/, ''], [/API key/, ''], [/Max number of files/, '10']].forEach(
       ([labelRe, expected]) => {
         const configInput = getByLabelText(labelRe) as HTMLInputElement;
         expect(configInput.value).toEqual(expected);
       }
     );
+
     [/Some object/, /Some other object/].forEach(labelRe => {
       const fieldCheckbox = getByLabelText(labelRe) as HTMLInputElement;
       expect(fieldCheckbox.checked).toBe(false);
@@ -79,6 +81,7 @@ describe('CloudinaryAppConfig', () => {
 
     const { getByLabelText } = renderComponent(sdk);
     await wait(() => getByLabelText(/Cloud name/));
+
     [
       [/Cloud name/, 'test-cloud'],
       [/API key/, 'test-api-key'],
@@ -87,6 +90,7 @@ describe('CloudinaryAppConfig', () => {
       const configInput = getByLabelText(labelRe as RegExp) as HTMLInputElement;
       expect(configInput.value).toEqual(expected);
     });
+    
     [[/Some object/, false], [/Some other object/, true]].forEach(([labelRe, expected]) => {
       const fieldCheckbox = getByLabelText(labelRe as RegExp) as HTMLInputElement;
       expect(fieldCheckbox.checked).toBe(expected);
