+++ marketplace/cloudinary-assets/src/components/cloudinaryField/cloudinaryField.tsx
@@ -3,7 +3,7 @@ import { Button } from '@contentful/forma-36-react-components';
 import { FieldExtensionSDK } from 'contentful-ui-extensions-sdk';
 import { CloudinaryResource } from '../../cloudinaryInterfaces';
 import { SortableComponent } from '../sortable/sortable';
-import { ExtensionParameters } from '../cloudinaryAppConfig/parameters';
+import { ExtensionParameters } from '../AppConfig/parameters';
 
 interface Props {
   sdk: FieldExtensionSDK;
